added safe delete service function

+++ lib/application/services/work_log_service.dart
@@ -265,15 +265,20 @@ class WorkLogService implements IWorkLogService {
   }
 
   @override
-  Future<Result<void, Failure>> bulkDeleteWorkLogs(List<int> ids) async {
+  Future<Result<void, Failure>> bulkDeleteWorkLogs({
+    required List<int> ids,
+    bool isSafeDelete = false,
+  }) async {
     try {
       List<WorkLog> workLogs = await _workLogRepository.getByIDList(ids);
       final syncedWorkLogs = workLogs.where(
         (workLog) => workLog.workLogState == WorkLogStateEnum.synced,
       );
 
-      for (final workLog in syncedWorkLogs) {
-        await _jiraWorkLogRepository.deleteJiraWorkLog(workLog);
+      if (!isSafeDelete) {
+        for (final workLog in syncedWorkLogs) {
+          await _jiraWorkLogRepository.deleteJiraWorkLog(workLog);
+        }
       }
 
       await _workLogRepository.bulkDeleteWorkLogs(ids);
